feat(frontend): SSR/SEO to forecast page
