+++ packages/sapience/src/lib/constants/focusAreas.ts
@@ -135,3 +135,16 @@ export const FOCUS_AREAS: FocusArea[] = [
 ];
 
 export const DEFAULT_FOCUS_AREA = FOCUS_AREAS[0];
+
+export const getFocusAreaMap = () => {
+  return new Map(
+    FOCUS_AREAS.map((area) => [
+      area.id,
+      {
+        iconSvg: area.iconSvg,
+        color: area.color,
+        name: area.name,
+      },
+    ])
+  );
+};
